requirements.txt
Django
djangorestframework
requests
drf-spectacular
psycopg[binary]==3.1.12
